pyproject.toml
[project]
name = "pptx_layout_engine"
version = "0.1.0"
description = "Generate PowerPoint presentations from YAML layout + content files"
authors = [{ name = "Tom Gillam", email = "tom.gillam@yahoo.com" }]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
  "python-pptx",
  "pyyaml"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
# Ensures non-Python files in resources/ are included
pptx_layout_engine = [
  "resources/slide_layouts/*.yml",
  "resources/shape_layouts/*.yml",
  "resources/powerpoint_templates/*.pptx"
]
